+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 3.16.0(49)
+### [3.16.0(49)](https://github.com/VakhoKontridze/VCore/releases/tag/3.16.0) — *2022 07 15*
 
 View and ViewControllers